Cleanup Universities (Non-Corporates group)

From non_corporates.do file. Source - https://sites.google.com/site/patentdataproject/Home/posts/namestandardizationroutinesuploaded

cockburn.replace.univ(x, ...)

Arguments

x

vector or table

...

Arguments passed on to harmonize.detect

patterns

Accepts both vector or table. If patterns it is table can also include replacements column.

patterns.col

If patterns is not a vector specifies which column to use. Default is 1.

patterns.as.codes

If set then use patterns as codes. In that case codes are ignored and there is no need to provide them. Default is FALSE.

patterns.codes.col

If patterns is table which column to use as codes column. Default is 2.

patterns.type

Specifies kind(s) of patterns. Default is "fixed" (calling codestri_replace_all_fixed). Other options are "begins", "ends" - which means that it should only match fixed pattern at the beginning of the string or at the and. Another possible value is "regex" (calling codestri_replace_all_regex). Another possible options are "exact" and "trim.exact" which matches a pattern string exactly or its trimmed version (i.e., white spaces are ignored at the beginning and at the end of the pattern)

patterns.type.col

Column in patterns table where you can specify types of patterns for each pattern. If set then patterns.type is ignored. Default is NULL

codes

If provided use it as codes. Should be the same length as patterns. Default is NULL.

codes.name

If provided use it as a name for codes column in results. It also will work as x.codes.col if x.codes.merge or x.codes.update.empyty are set. In case if x.codes.merge or x.codes.update.empyty are set and names(x) does not have column name codes.name it will reset both x.codes.merge and x.codes.update.empyty to FALSE and will (pre)append a new column codes.name.

codes.suffix

If codes.name is not provided use this suffix to append to x.col name or x.atomic.name if x is vector (see harmonize.x function).

codes.omitted.val

If x.rows is set. Use this value to fill the omitted rows. Default is vector of NAs of x length. When we update existing codes column (i.e., x.codes.col is set and valid) and the value is not set explicitly it resets to NULL which means that we use initial values for omitted rows.

codes.prepend

Whether to prepend codes column to the x. Default is FALSE. If set it will prepend codes column to x instead of append.

x.codes.col

If x is table, which column to use for making/merging/adding newly detected codes. Default is NULL. When it is not set explicitly and x.codes.merge or x.codes.update.empty are set it will first try to use codes.name if it is provided or if not will set the last column of x as x.codes.col.

x.codes.update.empty

If set then detect and add new codes only for records (rows) that were not yet coded (i.e., related codes are either "", NA or length == 0).

x.codes.merge

If set then merge (append or prepend) new codes to existing ones. It might turn the codes column to list of vectors.

x.codes.merge.prepend

If set then it will prepend new codes if not then it will append (default). It is ignored if x.codes.merge is not set.

return.only.codes

If set then just return codes vector. Default is FALSE. Basically when it is set it changes return.x.cols to 0.

return.only.first.detected.code

If TRUE then return only codes for the first detected pattern. If FALSE return list of vectors of all matched codes. Default is TRUE.

Value

Harmonized table