Harmonizes organizational names. Takes either vector or column in the table.

Harmonizes organizational names. Takes either vector or column in the table.

harmonize(
  x,
  procedures = harmonizer.default.procedures.table,
  progress = TRUE,
  progress.min = 10^5,
  progress.by = NA,
  progress.percent = 1,
  progress.message.use.names = TRUE,
  quite = FALSE,
  ...
)

Arguments

x

object (table)

procedures

Named list of procedures (closures) to apply to x. If we need to pass arguments to some of the procedures it can be done by specifying sub-list where the first element is procedure and the rest its arguments. Names of the list elements are used for progress messages. Procedures can also be passed as data.frame in which case it will be converted to list of procedures with harmonize.make.procedures.list (see its help for the correct format of data.frame with procedures). Default is harmonizer.default.procedures.table

progress

Show the progress? Default is TRUE

progress.min

The minimum number of rows the x should have for automatic progress estimation. If x has less rows no progress will be shown. Default is 10^5

progress.by

If set it will divide the x into chunk of this amount of rows. Default is NA.

progress.percent

Number of percents that represent one step in progress. Value should be between 0.1 and 50. Default is 1 which means it will try to chunk the x into 100 pieces.

progress.message.use.names

Should we use names from procedures list to report progress. Default is TRUE.

quite

Suppress all messages. Default is FALSE.

...

Arguments passed on to harmonize.x

inset.vector

a vector to inset. Optional. Default is NULL

x.col

vector of interest in x object

x.col.update

Update values in x.col column. Default is FALSE. If set inset.append, inset.name and inset.suffix are ignored. Also if set the default for return.x.cols.all will be set to TRUE.

x.rows

Logical vector to filter records of interest. Default is NULL which means do not filter records

x.atomic.name

If x is vector use this name for original column if it is in results. Default is "x". If x is table the name of x.col will be used.

inset.omitted.val

If x.rows is set merge these values to the results. It should be a vector of length 1 or nrow(x). If the value is NULL (default) then use values of x.col.

inset.append

If set then put inset.vector as the last instead of first vector/column. Default is FALSE.

inset.suffix

If inset.name is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already exists in return.x.cols it will add counter at the end to avoid variables with the same names.

inset.name

Use this name for the first column in results (harmonized names). Default is NULL, which means that either x.atomic.name if x is vector or original x.col name will be used with inset.suffix at the end.

return.x.cols

If x is table, set the columns to cbind to the result table. Default is cbind all but the original (unharmonized) column (x.col).

return.x.cols.all

Whether to bind all columns in x. Defaults depends on values of x.col.update and inset.append. If either is set then defaut values is TRUE otherwise FALSE. If set to TRUE by user the return.x.cols is ignored.

Value