This function is basically meant for coding names based on certain pattern

Optionally matches only at the beginning or at the end of the string.

harmonize.detect(
  x,
  patterns,
  patterns.col = 1,
  patterns.as.codes = FALSE,
  patterns.codes.col = 2,
  patterns.type = c("fixed", "begins", "begins.trimmed", "ends", "ends.trimmed",
    "regex", "exact", "exact.trimmed"),
  patterns.type.col = NULL,
  codes = NULL,
  codes.name = NA,
  codes.suffix = "coded",
  codes.omitted.val = NA,
  codes.prepend = FALSE,
  x.codes.col = NULL,
  x.codes.update.empty = FALSE,
  x.codes.merge = FALSE,
  x.codes.merge.prepend = FALSE,
  return.only.codes = FALSE,
  return.only.first.detected.code = FALSE,
  ...
)

Arguments

x

Vector or table to detect in.

patterns

Accepts both vector or table. If patterns it is table can also include replacements column.

patterns.col

If patterns is not a vector specifies which column to use. Default is 1.

patterns.as.codes

If set then use patterns as codes. In that case codes are ignored and there is no need to provide them. Default is FALSE.

patterns.codes.col

If patterns is table which column to use as codes column. Default is 2.

patterns.type

Specifies kind(s) of patterns. Default is "fixed" (calling codestri_replace_all_fixed). Other options are "begins", "ends" - which means that it should only match fixed pattern at the beginning of the string or at the and. Another possible value is "regex" (calling codestri_replace_all_regex). Another possible options are "exact" and "trim.exact" which matches a pattern string exactly or its trimmed version (i.e., white spaces are ignored at the beginning and at the end of the pattern)

patterns.type.col

Column in patterns table where you can specify types of patterns for each pattern. If set then patterns.type is ignored. Default is NULL

codes

If provided use it as codes. Should be the same length as patterns. Default is NULL.

codes.name

If provided use it as a name for codes column in results. It also will work as x.codes.col if x.codes.merge or x.codes.update.empyty are set. In case if x.codes.merge or x.codes.update.empyty are set and names(x) does not have column name codes.name it will reset both x.codes.merge and x.codes.update.empyty to FALSE and will (pre)append a new column codes.name.

codes.suffix

If codes.name is not provided use this suffix to append to x.col name or x.atomic.name if x is vector (see harmonize.x function).

codes.omitted.val

If x.rows is set. Use this value to fill the omitted rows. Default is vector of NAs of x length. When we update existing codes column (i.e., x.codes.col is set and valid) and the value is not set explicitly it resets to NULL which means that we use initial values for omitted rows.

codes.prepend

Whether to prepend codes column to the x. Default is FALSE. If set it will prepend codes column to x instead of append.

x.codes.col

If x is table, which column to use for making/merging/adding newly detected codes. Default is NULL. When it is not set explicitly and x.codes.merge or x.codes.update.empty are set it will first try to use codes.name if it is provided or if not will set the last column of x as x.codes.col.

x.codes.update.empty

If set then detect and add new codes only for records (rows) that were not yet coded (i.e., related codes are either "", NA or length == 0).

x.codes.merge

If set then merge (append or prepend) new codes to existing ones. It might turn the codes column to list of vectors.

x.codes.merge.prepend

If set then it will prepend new codes if not then it will append (default). It is ignored if x.codes.merge is not set.

return.only.codes

If set then just return codes vector. Default is FALSE. Basically when it is set it changes return.x.cols to 0.

return.only.first.detected.code

If TRUE then return only codes for the first detected pattern. If FALSE return list of vectors of all matched codes. Default is TRUE.

...

Arguments passed on to harmonize.x

inset.vector

a vector to inset. Optional. Default is NULL

x.col

vector of interest in x object

x.col.update

Update values in x.col column. Default is FALSE. If set inset.append, inset.name and inset.suffix are ignored. Also if set the default for return.x.cols.all will be set to TRUE.

x.rows

Logical vector to filter records of interest. Default is NULL which means do not filter records

x.atomic.name

If x is vector use this name for original column if it is in results. Default is "x". If x is table the name of x.col will be used.

inset.omitted.val

If x.rows is set merge these values to the results. It should be a vector of length 1 or nrow(x). If the value is NULL (default) then use values of x.col.

inset.append

If set then put inset.vector as the last instead of first vector/column. Default is FALSE.

inset.suffix

If inset.name is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already exists in return.x.cols it will add counter at the end to avoid variables with the same names.

inset.name

Use this name for the first column in results (harmonized names). Default is NULL, which means that either x.atomic.name if x is vector or original x.col name will be used with inset.suffix at the end.

return.x.cols

If x is table, set the columns to cbind to the result table. Default is cbind all but the original (unharmonized) column (x.col).

return.x.cols.all

Whether to bind all columns in x. Defaults depends on values of x.col.update and inset.append. If either is set then defaut values is TRUE otherwise FALSE. If set to TRUE by user the return.x.cols is ignored.

Value

The updated x table with codes column or just codes if return.only.codes is set.