Makes list of procedures calls from table.

Table should have at least two columns - messages and fuctions calls. Each function call should be a string of the following format "'function.name', arg1 = val1, arg2 = val2" (same as arguments for do.call function).

harmonize.make.procedures.list(
  procedures.table,
  message.field = "message",
  function.call.field = "function.call",
  no.field = "no",
  remove.comments = TRUE,
  sort.by.no.field = TRUE,
  comments = c("#", "-", "")
)

Arguments

procedures.table

Table to use

message.field

name of the column with messages that will be displayed when each call is executed

function.call.field

name of the column where function (harmonization procedures) calls are listed.

no.field

name of the column where the number of procedure is specified. Also this field indicates if the row in the table is just a comment in which case it will be removed if remove.comments is set (which is set by default)

remove.comments

Whether to remove comments.

Value

List of named function calls. Names are messages.