Uppercases vector of interest in the object (table)

Uppercases vector of interest in the object (table)

harmonize.toupper(data, ...)

Arguments

data
data

...
Arguments passed on to inset_vector

vector

a vector to inset. Optional. Default is NULL

col

vector of interest in data object

update

Update values in col column. Default is FALSE. If set append, name and name_suffix are ignored. Also if set the default for select_all_data_cols will be set to TRUE.

rows

Logical vector to filter records of interest. Default is NULL which means do not filter records

ommitted_rows_values

If rows is set merge these values to the results. It should be a vector of length 1 or nrow(data). If the value is NULL (default) then use original values of col.

append

If set then put vector as the last instead of first vector/column. Default is FALSE.

name_suffix

If name is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already edataists in select_data_cols it will add counter at the end to avoid variables with the same names.

name

Use this name for the first column in results (harmonized names). Default is NULL, which means that either name_if_data_atomic if data is vector or original col name will be used with name_suffix at the end.

name_if_data_atomic

If data is vector use this name for original column if it is in results. Default is "data". If data is table the name of col will be used.

select_data_cols

If data is table, set the columns to cbind to the result table. Default is cbind all but the original (unharmonized) column (col).

select_all_data_cols

Whether to bind all columns in data. Defaults depends on values of update and append. If either is set then defaut values is TRUE otherwise FALSE. If set to TRUE by user the select_data_cols is ignored.

Value

updated data (as data.table)