Gets a vector to harmonize and puts it back.

The function harmonize.x basically works as two functions depending whether the second optional parameter inset.vector is provided. If inset.vector is not provided the function returns a vector that we want to process (harmonize) from object x and inset it back to the original object later. If inset.vector (harmonized vector) is provided the function returns updated x.

harmonize.x(
  x,
  inset.vector = NULL,
  x.col = 1,
  x.col.update = FALSE,
  x.rows = NULL,
  x.atomic.name = "x",
  inset.omitted.val = NULL,
  inset.append = FALSE,
  inset.suffix = "harmonized",
  inset.name = NA,
  return.x.cols = -ifelse(is.numeric(x.col), x.col, match(x.col, names(x))),
  return.x.cols.all = inset.append | x.col.update
)

Arguments

x
an object

inset.vector
a vector to inset. Optional. Default is NULL

x.col
vector of interest in x object

x.col.update
Update values in x.col column. Default is FALSE. If set inset.append, inset.name and inset.suffix are ignored. Also if set the default for return.x.cols.all will be set to TRUE.

x.rows
Logical vector to filter records of interest. Default is NULL which means do not filter records

x.atomic.name
If x is vector use this name for original column if it is in results. Default is "x". If x is table the name of x.col will be used.

inset.omitted.val
If x.rows is set merge these values to the results. It should be a vector of length 1 or nrow(x). If the value is NULL (default) then use values of x.col.

inset.append
If set then put inset.vector as the last instead of first vector/column. Default is FALSE.

inset.suffix
If inset.name is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already exists in return.x.cols it will add counter at the end to avoid variables with the same names.

inset.name
Use this name for the first column in results (harmonized names). Default is NULL, which means that either x.atomic.name if x is vector or original x.col name will be used with inset.suffix at the end.

return.x.cols
If x is table, set the columns to cbind to the result table. Default is cbind all but the original (unharmonized) column (x.col).

return.x.cols.all
Whether to bind all columns in x. Defaults depends on values of x.col.update and inset.append. If either is set then defaut values is TRUE otherwise FALSE. If set to TRUE by user the return.x.cols is ignored.

Value

Vector or data.table