Harmonizes strings using exact procedures described in Magerman et al. 2009.

Harmonizes strings using exact procedures described in Magerman et al. 2009.

harmonize_magerman(
  x,
  magerman.procedures = magerman.procedures.table,
  detect.legal.form = FALSE,
  return.x.before.common.words.removal = FALSE,
  return.x.cols.all = FALSE,
  ...
)

Arguments

x

table or vector

magerman.procedures

Named list of procedures (closures) to apply to x. If we need to pass arguments to some of the procedures it can be done by specifying sub-list where the first element is procedure and the rest its arguments. Names of the list elements are used for progress messages. Procedures can also be passed as data.frame in which case it will be converted to list of procedures with harmonize.make.procedures.list (see its help for the correct format of data.frame with procedures). Default is magerman.procedures.table.

detect.legal.form

Whether to detect legal forms. Default is FALSE

return.x.before.common.words.removal

Whether to save harmonized column before common.words.removal procedure. Default is FALSE

return.x.cols.all

Whether to return initial column in x. Default is FALSE

...

Arguments passed on to harmonize

procedures

Named list of procedures (closures) to apply to x. If we need to pass arguments to some of the procedures it can be done by specifying sub-list where the first element is procedure and the rest its arguments. Names of the list elements are used for progress messages. Procedures can also be passed as data.frame in which case it will be converted to list of procedures with harmonize.make.procedures.list (see its help for the correct format of data.frame with procedures). Default is harmonizer.default.procedures.table

progress

Show the progress? Default is TRUE

progress.min

The minimum number of rows the x should have for automatic progress estimation. If x has less rows no progress will be shown. Default is 10^5

progress.by

If set it will divide the x into chunk of this amount of rows. Default is NA.

progress.percent

Number of percents that represent one step in progress. Value should be between 0.1 and 50. Default is 1 which means it will try to chunk the x into 100 pieces.

progress.message.use.names

Should we use names from procedures list to report progress. Default is TRUE.

quite

Suppress all messages. Default is FALSE.

Value

Harmonized table

References

Magerman et al., 2006 - Data Production Methods for Harmonized Patent Statistics: Patentee Name Harmonization