This function provides options for harmonizer functions that controls how the input data object, intermediate and end results are handled.

It solves problem of passing options parameters to some arbitrary harmonizing function that are also passed with ... (in which case there will be an error that parameters have duplicated names). Using this function allows not to pass ... explicitly down to calling g stack and get the options values updated by explicit parameters at any location in the stack.

harmonize_options(
  col = 1,
  rows = NULL,
  ommitted_rows_values = NULL,
  placement = c("replace_col", "prepend_to_data", "append_to_data", "prepend_to_col",
    "append_to_col", "ignore"),
  name = NA,
  name_if_input_atomic = "names",
  name_suffix = "_harmonized",
  append_copy = FALSE,
  append_copy_name_format = "%name_harmonizing_%number_%procedure",
  select_cols = NA
)

Arguments

col
Column of interest in the data object. The one we need to harmonize

rows
Logical vector to filter records of interest. Default is NULL which means do not filter records

ommitted_rows_values
If rows is set merge these values to the results. It should be a vector of length 1 or nrow(data). If the value is NULL (default) then use original values of col`

placement
Where to inset retults (harmonized vector) in the data object. Default options is "replace_col" which overwrides the col in data with results.

name
Use this name for the first column in results (harmonized names). Default is NULL, which means that either name_if_data_atomic if data is vector or original col name will be used with name_suffix at the end.

name_if_input_atomic
If data is vector use this name for original column if it is in results. Default is "data". If data is table the name of col will be used.

name_suffix
If name is not set the use this as suffix (default is "harmonized"). If the name with the same suffix already edataists in select_data_cols it will add counter at the end to avoid variables with the same names.

append_copy
Whether to append a copy of result vector to data object

append_copy_name_format
How the append copy wiil be named

select_cols
If data` object is table, set the columns to cbind to the result table. Default is cbind all but the original (unharmonized) column (col).

Value

list of options updated with arguments from harmonizer functions calls stack up to .GlobalEnv or 'harmonize' call.