Replaces comma period irregularities

Replaces comma period irregularities

magerman_replace_comma_period_irregularities_all(x, ...)

Arguments

x

table

...

Arguments passed on to harmonize_options

col

Column of interest (the one we need to harmonize) in the x object (if it is data.frame like).

rows

Logical vector to filter records of interest. Default is NULL which means do not filter records.

omitted_rows_value

If rows parameter is set then merge omitted_rows_value with the results (filtered by rows). Either single string or a character vector of length nrow(x). If NULL (the default) then original values of col are merged with results.

output_placement

Where to inset retults (harmonized vector) in the x object. Default options is 'replace_col' which overwrides the col in x with results. Other options:

'omit' :: do not write results back to table (usually used when append_output_copy is set for temporary values)

'prepend_to_col' :: prepend to col

'append_to_col' :: append to col

'prepend_to_x' :: prepend to x data.frame like object

'append_to_x' :: append to x data.frame like object

atomic_x_col_name

If x is vector use this name for original column if it is in results. Default is "x". If x is table the name of col will be used.

output_name

Use this name for the column with results (harmonized values). Parts in curly brakeds are substitute strings. Options for substitutions are:

append_output_copy

Whether to append a copy of result vector to x object

output_copy_col_name

How the append copy wiil be named

Value

Harmonized table