Replaces sgml characters. Accept both vector and table and return either vector or a table

Replaces sgml characters. Accept both vector and table and return either vector or a table

magerman_replace_sgml_characters(x, ...)

Arguments

x
table

...
Arguments passed on to harmonize.replace

patterns

Accepts both vector or table. If patterns it is table can also include replacements column.

patterns.col

If patterns is not a vector which column to use. Default is 1.

patterns.mode

Mode of matching. Could be one of c("all", "first", "last"). The default is "all" (it is 2x faster than "first" and "last" because of handy stri_replace_all_* functions). Also possible to pass a vector (same length as patterns)

patterns.mode.col

Column in patterns table with the mode of matching

patterns.type

Kind of pattern. Default is "fixed" (calling codestri_replace_all_fixed). Other options are "begins", "ends" - which means that it should only match fixed pattern at the beginning of the string or at the and. Another possible value is "regex" (calling codestri_replace_all_regex)

patterns.type.col

Column with the type of pattern in case when patterns should have different types

patterns.replacements.col

If patterns is not a vector and includes replacements which column to use for replacements. Default is 2.

replacements

If patterns does not have column with replacements provide it here.

Value

Harmonized table